Move ftp classes to separate package
