+++ src/main/java/uk/gov/hmcts/reform/sendletter/tasks/StaleLettersTask.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Component;
 import uk.gov.hmcts.reform.sendletter.entity.LetterRepository;
 import uk.gov.hmcts.reform.sendletter.entity.LetterStatus;
 import uk.gov.hmcts.reform.sendletter.logging.AppInsights;
-import uk.gov.hmcts.reform.sendletter.services.FtpAvailabilityChecker;
+import uk.gov.hmcts.reform.sendletter.services.ftp.FtpAvailabilityChecker;
 
 import java.sql.Timestamp;
 import java.time.LocalDateTime;
